fix: delete v3 version
